requirements.txt
bottle==0.12.25
packaging==24.1
proxy-tools==0.1.0
pycairo==1.26.1
pygobject==3.48.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.14
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.14
pywebview==5.1
QtPy==2.4.1
typing-extensions==4.12.2
